text-input: Synchronize set_cursor_rectangle with the wl_surface
Currently, the after-effects of the set_cursor_rectangle request on
the input method UI and the client-side redraws are not guaranteed
to be atomic.

Add the requirement for compositors handling the newer version of
the protocol to make the set_cursor_rectangle request visually
effective with the same wl_surface.commit.

Signed-off-by: Tadeo Kondrak <me@tadeo.ca>